pm/amd-pstate: Put long links and paths in a new line
Doxygen does not wrap links and paths that continue from the previous
line. Work around the limitation by putting the links and paths in a
separate line.

Since Github renders a tabbed list data with an empty line on top as a
codeblock, use "  \n" instead to keep both Github and Doxygen happy.

Reviewed-by: Dhananjay Ugwekar <dhananjay.ugwekar@amd.com>
Signed-off-by: K Prateek Nayak <kprateek.nayak@amd.com>